Rename dialer to client
